Replace use of size_t and %zu with unsigned long and %lu due to some poor implementations of %zu.

+++ iocccsize.h
@@ -42,14 +42,14 @@ extern "C" {
 #endif
 
 typedef struct {
-	size_t rule_2a_size;
-	size_t rule_2b_size;
-	size_t keywords;	/* keyword count - for -v mode */
-	size_t nul;		/* Count NUL bytes seen. */
-	size_t high_bit;	/* Count non-ASCII high-bit set bytes. */
-	size_t bad_trigraph;	/* Count unknown trigraphs seen. */
-	size_t ungetc_error;	/* Count ungetc depth exceeded. */
-	size_t word_overflow;	/* Count word buffer overflows. */
+	unsigned long rule_2a_size;
+	unsigned long rule_2b_size;
+	unsigned long keywords;		/* keyword count - for -v mode */
+	unsigned long nul;		/* Count NUL bytes seen. */
+	unsigned long high_bit;		/* Count non-ASCII high-bit set bytes. */
+	unsigned long bad_trigraph;	/* Count unknown trigraphs seen. */
+	unsigned long ungetc_error;	/* Count ungetc depth exceeded. */
+	unsigned long word_overflow;	/* Count word buffer overflows. */
 } RuleCount;
 
 /**
